Fix for warning when building signal.c, fs-rename.c, osword.c
Forgot to include string.h in signal.
Forgot to include fs-errors.h in fs-rename.c.
Forgot to cast in kernel-osword.c.

+++ clib/kernel/kernel-osword.c
@@ -19,7 +19,7 @@ int _kernel_osword(int op, int *data)
     _kernel_swi_regs regs;
 
     regs.r[0] = op;
-    regs.r[1] = data;
+    regs.r[1] = (int)data;
 
     err = _kernel_swi(OS_Word, &regs, &regs);
     if (err)
